made a new view model
